*: Rename aggregate-co -> clusteroperator
